Refactor: Move core model to domain model
