+++ app/src/main/java/com/severo/marvel/presentation/characters/adapters/CharactersAdapter.kt
@@ -3,7 +3,7 @@ package com.severo.marvel.presentation.characters.adapters
 import android.view.ViewGroup
 import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DiffUtil
-import com.severo.core.model.Character
+import com.severo.core.domain.model.Character
 import com.severo.marvel.framework.imageloader.ImageLoader
 import com.severo.marvel.util.OnCharacterItemClick
