+++ app/src/main/res/values/colors.xml
@@ -8,10 +8,12 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
 
-    <color name="medium_grey">#D1D1D1</color>
-    <color name="black_800">#484848</color>
+    <color name="red_400">#FF5F52</color>
     <color name="red_800">#C62828</color>
-
+    <color name="red_900">#8E0000</color>
+    <color name="black_900">#212121</color>
+    <color name="black_800">#484848</color>
+    <color name="medium_grey">#D1D1D1</color>
     <color name="checked_chip">#20C62828</color>
     <color name="unchecked_chip">@color/medium_grey</color>
     <color name="checked_chip_text">@color/red_800</color>
