+++ config/detekt/detekt.gradle
@@ -1,5 +1,35 @@
 apply plugin: "io.gitlab.arturbosch.detekt"
+
+project.afterEvaluate {
+    tasks.named("preBuild") {
+        dependsOn("detekt")
+    }
+}
+
 detekt {
-    toolVersion = "1.23.3"
-    config.setFrom(file("$rootDir/config/detekt/detekt.yml"))
+    toolVersion = "1.18.0"
+    def inputDirFiles = []
+    rootProject.subprojects.each { module ->
+        if (module.name != "testing") {
+            inputDirFiles << "$module.projectDir/src/main/java"
+        }
+    }
+    source = files(inputDirFiles)
+    config = files("$rootDir/config/detekt/detekt.yml")
+    autoCorrect = true
+
+    reports {
+        xml {
+            enabled = true
+            destination = file("$buildDir/reports/detekt/detekt.xml")
+        }
+        html {
+            enabled = true
+            destination = file("$buildDir/reports/detekt/detekt.html")
+        }
+        txt {
+            enabled = true
+            destination = file("$buildDir/reports/detekt/detekt.txt")
+        }
+    }
 }
