+++ core/src/main/java/com/severo/core/usecase/CheckFavoriteUseCase.kt
@@ -6,7 +6,6 @@ import com.severo.core.usecase.base.ResultStatus
 import com.severo.core.usecase.base.UseCase
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
-import javax.inject.Inject
 
 interface CheckFavoriteUseCase {
 
@@ -15,10 +14,10 @@ interface CheckFavoriteUseCase {
     data class Params(val characterId: Int)
 }
 
-class CheckFavoriteUseCaseImpl @Inject constructor(
+class CheckFavoriteUseCaseImpl(
     private val favoritesRepository: FavoritesRepository,
     private val dispatchers: CoroutinesDispatchers
-) : UseCase<CheckFavoriteUseCase.Params, Boolean>(), CheckFavoriteUseCase {
+): UseCase<CheckFavoriteUseCase.Params, Boolean>(), CheckFavoriteUseCase {
 
     override suspend fun doWork(params: CheckFavoriteUseCase.Params): ResultStatus<Boolean> {
         return withContext(dispatchers.io()) {
